@@ -3,7 +3,7 @@
     "description": "A CLI tool that converts PHP files to use the PHP 5.4+ shorthand array syntax.",
     "license": "MIT",
     "require": {
-        "php": ">=5.4.0",
+        "php": ">=7.0",
         "symfony/console": "*"
     },
     "require-dev": {
